requirements.txt
python-dotenv==1.0.0
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.18.1
colorama==0.4.6
tqdm==4.66.2